Added compatibility of floor_ste and ceil_ste with Pytorch JIT and moved them in appropriate files
